+++ R/prob-roc_auc.R
@@ -60,11 +60,19 @@
 #'
 #' @author Max Kuhn
 #'
-#' @template examples-prob
+#' @template examples-binary-prob
+#' @template examples-multiclass-prob
 #' @examples
-#' # passing options via a list and _not_ `...`
-#' roc_auc(two_class_example, truth = truth, Class1,
-#'         options = list(smooth = TRUE))
+#' # ---------------------------------------------------------------------------
+#' # Options for `pROC::roc()`
+#'
+#' # Pass options via a named list and not through `...`!
+#' roc_auc(
+#'   two_class_example,
+#'   truth = truth,
+#'   Class1,
+#'   options = list(smooth = TRUE)
+#' )
 #'
 #' @export
 roc_auc <- function(data, ...) {