+++ R/metrics.R
@@ -167,7 +167,6 @@ metrics.data.frame <- function(data, truth, estimate, ...,
 #' probability columns) as bare column names or `tidyselect` selectors to `...`.
 #'
 #' @examples
-#'
 #' library(dplyr)
 #'
 #' # Multiple regression metrics
@@ -189,7 +188,8 @@ metrics.data.frame <- function(data, truth, estimate, ...,
 #' # If you need to set options for certain metrics,
 #' # do so by wrapping the metric and setting the options inside the wrapper,
 #' # passing along truth and estimate as quoted arguments.
-#' # Then add on the function class of the underlying wrapped function.
+#' # Then add on the function class of the underlying wrapped function,
+#' # and the direction of optimization.
 #' ccc_with_bias <- function(data, truth, estimate, na_rm = TRUE, ...) {
 #'   ccc(
 #'     data = data,
@@ -202,8 +202,10 @@ metrics.data.frame <- function(data, truth, estimate, ...,
 #'   )
 #' }
 #'
-#' # Add on the underlying function class (here, "numeric_metric")
+#' # Add on the underlying function class (here, "numeric_metric"), and the
+#' # direction to optimize the metric
 #' class(ccc_with_bias) <- class(ccc)
+#' attr(ccc_with_bias, "direction") <- attr(ccc, "direction")
 #'
 #' multi_metric2 <- metric_set(rmse, rsq, ccc_with_bias)
 #'
@@ -349,11 +351,13 @@ make_prob_class_metric_function <- function(fns) {
     class(metric_function)
   )
 
+  attr(metric_function, "metrics") <- fns
+
   metric_function
 }
 
 make_numeric_metric_function <- function(fns) {
-  numeric_metric_function <- function(data, truth, estimate, na_rm = TRUE, ...) {
+  metric_function <- function(data, truth, estimate, na_rm = TRUE, ...) {
 
     # Construct common argument set for each metric call
     # Doing this dynamically inside the generated function means
@@ -381,12 +385,14 @@ make_numeric_metric_function <- function(fns) {
     bind_rows(metric_list)
   }
 
-  class(numeric_metric_function) <- c(
+  class(metric_function) <- c(
     "numeric_metric_set",
-    class(numeric_metric_function)
+    class(metric_function)
   )
 
-  numeric_metric_function
+  attr(metric_function, "metrics") <- fns
+
+  metric_function
 }
 
 validate_not_empty <- function(x) {