+++ NEWS.md
@@ -1,3 +1,5 @@
+# yardstick development
+
 # yardstick 0.0.3
 
 ## New metrics and functionality